+++ x/wasm/artifacts/v152/api/callbacks_cgo.go
@@ -1,7 +1,7 @@
 package api
 
 /*
-#include "bindings.h"
+#include "bindings152.h"
 #include <stdio.h>
 
 // imports (db)
