+++ x/wasm/artifacts/v152/api/lib.go
@@ -1,7 +1,7 @@
 package api
 
 // #include <stdlib.h>
-// #include "bindings.h"
+// #include "bindings152.h"
 import "C"
 
 import (
