+++ x/wasm/artifacts/v152/api/version.go
@@ -1,7 +1,7 @@
 package api
 
 /*
-#include "bindings.h"
+#include "bindings152.h"
 */
 import "C"
